Create more files to separe responsabilities
