+++ package.json
@@ -24,7 +24,7 @@
   "scripts": {
     "test": "jest",
     "test:src": "jest src",
-    "test:integration": "jest __tests__",
+    "test:cli": "jest __tests__",
     "start": "node lib/cli/index.js",
     "build": "yarn babel src --extensions \".js,.ts\" --out-dir lib --copy-files --no-copy-ignored",
     "dev": "ts-node -r tsconfig-paths/register src/cli/index.ts",
@@ -64,6 +64,7 @@
     "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-standard": "^5.0.0",
     "jest": "^27.0.4",
+    "mock-fs": "^5.0.0",
     "prettier": "^2.3.1",
     "rimraf": "^3.0.2",
     "ts-node-dev": "^1.1.6",
