+++ src/core/commands/CreateCommand/Defaults.ts
@@ -1,7 +1,7 @@
-export const DefaultOptions = {
+export const CREATE_COMMAND_DEFAULT_OPTIONS = {
   name: '',
   templatesFolder: '__file-templates__',
-  encoding: 'utf-8',
+  encoding: 'utf-8' as BufferEncoding,
   replaceNames: undefined,
   replaceContent: undefined,
   brackets: true,
